Move DeclEnum-specific tests to tests.test_declenum
